requirements.txt
# Telegram
pyrogram==2.0.106
TgCrypto==1.2.5

# Web framework – fully async (replaces Flask)
aiohttp==3.9.5
aiohttp-jinja2==1.6
jinja2==3.1.4

# Database
motor==3.3.2
pymongo==4.6.1

# Utilities
python-dotenv==1.0.0
cryptography==41.0.7
